Feature/api horse jockey registration

+++ src/main/java/com/swp391/horseracing/dto/jockey/request/JockeyCreationRequest.java
@@ -21,10 +21,10 @@ public class JockeyCreationRequest {
     UUID userId;
 
     @Positive(message = "Height must be positive")
-    BigDecimal height;
+    float height;
 
     @Positive(message = "Weight must be positive")
-    BigDecimal weight;
+    float weight;
 
     @NotNull(message = "Experience year is required")
     @Min(value = 0, message = "Experience year must be at least 0")
